refactor : 검증 로직을 서비스에 위임
