+++ backend/smody/src/main/java/com/woowacourse/smody/controller/MemberController.java
@@ -53,9 +53,4 @@ public class MemberController {
         memberService.withdraw(tokenPayload);
         return ResponseEntity.noContent().build();
     }
-    
-    @GetMapping("/auth")
-    public ResponseEntity<ValidAuthResponse> isValid(@PreLoginMember PreTokenPayLoad preTokenPayLoad) {
-        return ResponseEntity.ok(memberService.isValidAuth(preTokenPayLoad));
-    }
 }
