+++ backend/smody/src/main/java/com/woowacourse/smody/service/OauthService.java
@@ -2,9 +2,7 @@ package com.woowacourse.smody.service;
 
 import com.woowacourse.smody.auth.JwtTokenProvider;
 import com.woowacourse.smody.domain.Member;
-import com.woowacourse.smody.dto.LoginRequest;
-import com.woowacourse.smody.dto.LoginResponse;
-import com.woowacourse.smody.dto.TokenPayload;
+import com.woowacourse.smody.dto.*;
 import com.woowacourse.smody.repository.MemberRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -29,5 +27,10 @@ public class OauthService {
 
     private String createToken(Member member) {
         return jwtTokenProvider.createToken(new TokenPayload(member.getId()));
+    }
+
+
+    public ValidAuthResponse isValidAuth(PreTokenPayLoad preTokenPayLoad) {
+        return new ValidAuthResponse(jwtTokenProvider.validateToken(preTokenPayLoad.getToken()));
     }
 }
