+++ backend/smody/src/test/java/com/woowacourse/smody/controller/MemberControllerTest.java
@@ -117,27 +117,4 @@ public class MemberControllerTest extends ControllerTest {
                         preprocessResponse(prettyPrint())
                 ));
     }
-
-    @DisplayName("나의 토큰의 유효성을 조회한다.")
-    @Test
-    void isValidAuth() throws Exception {
-        // given
-        String token = jwtTokenProvider.createToken(new TokenPayload(1L));
-        ValidAuthResponse validAuthResponse = new ValidAuthResponse(true);
-        given(memberService.isValidAuth(any(PreTokenPayLoad.class)))
-                .willReturn(validAuthResponse);
-
-        // when
-        ResultActions result = mockMvc.perform(get("/members/auth")
-                .header("Authorization", "Bearer " + token));
-
-        // then
-        result.andExpect(status().isOk())
-                .andExpect(content().json(objectMapper.writeValueAsString(validAuthResponse)))
-                .andDo(document("validate_auth", HOST_INFO,
-                        preprocessResponse(prettyPrint()),
-                        responseFields(
-                                fieldWithPath("isValid").type(JsonFieldType.BOOLEAN).description("토큰 유효 여부")
-                        )));
-    }
 }
