+++ backend/smody/src/test/java/com/woowacourse/smody/controller/OauthControllerTest.java
@@ -11,8 +11,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import com.woowacourse.smody.dto.LoginRequest;
-import com.woowacourse.smody.dto.LoginResponse;
+import com.woowacourse.smody.dto.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.restdocs.payload.JsonFieldType;
@@ -42,6 +41,29 @@ class OauthControllerTest extends ControllerTest {
                         responseFields(
                                 fieldWithPath("accessToken").type(JsonFieldType.STRING).description("엑세스 토큰"),
                                 fieldWithPath("isNewMember").type(JsonFieldType.BOOLEAN).description("신규 회원인지 여부")
+                        )));
+    }
+
+    @DisplayName("나의 토큰의 유효성을 조회한다.")
+    @Test
+    void isValidAuth() throws Exception {
+        // given
+        String token = jwtTokenProvider.createToken(new TokenPayload(1L));
+        ValidAuthResponse validAuthResponse = new ValidAuthResponse(true);
+        given(oauthService.isValidAuth(any(PreTokenPayLoad.class)))
+                .willReturn(validAuthResponse);
+
+        // when
+        ResultActions result = mockMvc.perform(get("/oauth/auth")
+                .header("Authorization", "Bearer " + token));
+
+        // then
+        result.andExpect(status().isOk())
+                .andExpect(content().json(objectMapper.writeValueAsString(validAuthResponse)))
+                .andDo(document("validate_auth", HOST_INFO,
+                        preprocessResponse(prettyPrint()),
+                        responseFields(
+                                fieldWithPath("isValid").type(JsonFieldType.BOOLEAN).description("토큰 유효 여부")
                         )));
     }
 }
